test: fix Integer(128, literal, party) reading stack garbage
BitVec_T<Wire>::BitVec_T(size_t width, T value, int party) (in
emp-tool/circuits/bitvec.hpp) calls
`bits_to_bools(tmp, &value, width)` to feed the input. With a plain
literal like `0` or `2`, T deduces to `int` (4 bytes), but
`bits_to_bools` reads ⌈width/8⌉ = 16 bytes for width=128 — so 12
bytes of stack garbage past the int get fed into the circuit as the
high bits of the input.

The bug was silent on macOS arm64 (Release-build stack frame happened
to be zero-filled past the int), but bit on Linux x86_64 in CI: the
test_aes ciphertext came out non-FIPS-vector and the memcmp errored.

Fix: cast the literal to __int128 so T deduces to a 16-byte type and
the full feed buffer is initialized. The BitVec_T (and SignedInt_T)
ctor's SFINAE explicitly accepts __int128 / __uint128_t.

- test_aes.cpp: plaintext / key / ciphertext (all width=128).
- test_circuit_file.cpp: a / b / c (all width=128). The output was
  only printed, not asserted, so the bug was silently producing
  wrong AES outputs there too — fixing for hygiene.

Verified: ctest still green locally on macOS; the cast also closes
the path on hosts where the stack-garbage isn't zero.

+++ test/test_aes.cpp
@@ -9,9 +9,14 @@ int main(int argc, char** argv) {
 	NetIO io(party == ALICE ? nullptr : "127.0.0.1", port);
 	setup_semi_honest(&io, party);
 
-	Integer plaintext(128, 0, ALICE);
-	Integer key(128, 0, BOB);
-	Integer ciphertext(128, 0, PUBLIC);
+	// Cast the literal to __int128 so the (size_t, T, int) ctor copies
+	// a full 16 bytes — passing a plain `0` (int) makes BitVec_T's
+	// `bits_to_bools(tmp, &value, 128)` read 12 bytes of stack garbage
+	// past the int, producing party-specific non-zero inputs that just
+	// happen to be all-zero on some hosts.
+	Integer plaintext(128, (__int128)0, ALICE);
+	Integer key(128, (__int128)0, BOB);
+	Integer ciphertext(128, (__int128)0, PUBLIC);
 
 	AES_Calculator aes;
 	aes.encrypt_with_key(plaintext.bits.data(),
